requirements.txt
ultralytics>=8.4.46
opencv-python>=4.9.0
numpy>=1.24
pandas>=2.0
pyfirmata==1.1.0
Pillow>=10.0
PyYAML>=6.0
